+++ src/shared-resources/components/Tooltip/Tooltip.tsx
@@ -1,27 +1,11 @@
 import React, { FC, ReactElement } from 'react';
 import RCTooltip from 'rc-tooltip';
 import 'rc-tooltip/assets/bootstrap.css';
-
-export enum TooltipTrigger {
-  HOVER = 'hover',
-  CLICK = 'click',
-  FOCUS = 'focus',
-}
-
-export enum TooltipPlacement {
-  TOP = 'top',
-  LEFT = 'left',
-  RIGHT = 'right',
-  BOTTOM = 'bottom',
-  TOP_LEFT = 'topLeft',
-  TOP_RIGHT = 'topRight',
-  BOTTOM_LEFT = 'bottomLeft',
-  BOTTOM_RIGHT = 'bottomRight',
-  LEFT_TOP = 'leftTop',
-  LEFT_BOTTOM = 'leftBottom',
-  RIGHT_TOP = 'rightTop',
-  RIGHT_BOTTOM = 'rightBottom',
-}
+// types
+import {
+  TooltipPlacement,
+  TooltipTrigger,
+} from 'shared-resources/types/Tooltip.type';
 
 interface TooltipProps {
   children: ReactElement;
