+++ src/shared-resources/types/Tooltip.type.ts
@@ -1,0 +1,20 @@
+export enum TooltipTrigger {
+  HOVER = 'hover',
+  CLICK = 'click',
+  FOCUS = 'focus',
+}
+
+export enum TooltipPlacement {
+  TOP = 'top',
+  LEFT = 'left',
+  RIGHT = 'right',
+  BOTTOM = 'bottom',
+  TOP_LEFT = 'topLeft',
+  TOP_RIGHT = 'topRight',
+  BOTTOM_LEFT = 'bottomLeft',
+  BOTTOM_RIGHT = 'bottomRight',
+  LEFT_TOP = 'leftTop',
+  LEFT_BOTTOM = 'leftBottom',
+  RIGHT_TOP = 'rightTop',
+  RIGHT_BOTTOM = 'rightBottom',
+}
